web/backend/requirements.txt
# FastAPI Backend Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# CORS
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# Core OCR dependencies (from main requirements.txt)
python-docx==1.1.0
Pillow==10.1.0
PyPDF2==3.0.1
pdf2image==1.16.3

# LLM API clients
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2
dashscope==1.14.1

# Formula conversion
latex2mathml==3.77.0

# Configuration
PyYAML==6.0.1
requests==2.31.0

# Logging
colorlog==6.8.0

# Image processing
opencv-python==4.8.1.78
numpy==1.26.2
